+++ bin/check-es-cluster-status.rb
@@ -84,13 +84,13 @@ class ESClusterStatus < Sensu::Plugin::Check::CLI
          long: '--https'
 
   option :yellow_only,
-         description: 'If you would only like to be alerted for a yellow (or green) state'
-         long: '--yellow_only'
+         description: 'If you would only like to be alerted for a yellow (or green) state',
+         long: '--yellow_only',
          boolean: true
 
   option :red_only,
-         description: 'If you would only like to be alerted for a red (or green) state'
-         long: '--red_only'
+         description: 'If you would only like to be alerted for a red (or green) state',
+         long: '--red_only',
          boolean: true
 
   def get_es_resource(resource)
@@ -149,9 +149,17 @@ class ESClusterStatus < Sensu::Plugin::Check::CLI
       when 'green'
         ok 'Cluster is green'
       when 'yellow'
-        options[:red_only] ? ok : warning 'Cluster state is Yellow'
+        if !options[:red_only]
+          warning 'Cluster state is Yellow'
+        else
+          ok 'Not alerting on yellow'
+        end
       when 'red'
-        options[:yellow_only] ? ok : critical 'Cluster state is Red'
+        if !options[:yellow_only]
+          critical 'Cluster state is Red'
+        else
+          ok 'Not alerting on red'
+        end
       end
     else
       ok 'Not the master'